yocto: amd: make changes to AMD SPI SDK for meta-baldeagle
There was an issue reported where the SPI kernel thread would
take close to 100% of CPU cycles even when there was no SPI
transaction actually taking place. The root cause was that
the kernel thread was contiously running in a loop without
ever going into sleep.

To fix the issue, the SPI thread is put into a wait queue untill
there was one or more SPI messages to be processed. This will
ensure that the thread only comes consumes cycles while processing
SPI transactions. Rest of the time, it will just sleep.

Upstream Status: None