busybox: fix compile failures on terminated builds
The do_compile step for busybox does configuration stuff
as well which can not be avoided. While doing so it updates
the .config that is used to specify the package configurations
and pass those on. If the do_compile step is terminated in
between due to any reason it can end up in a state where the
busybox package gets misconfigured.
In order to avoid this issue we archive the .config generated
by the do_configure step and start off fresh everytime for the
do_compile.

Signed-off-by: Awais Belal <awais_belal@mentor.com>