archive-release: Fix bashism in git_tar
Override the git_tar routine from the upstream bb file.
There is an errant bashism (ie comparison in an if statement
using "==" rather than "=") which fails on systems using
dash as /bin/sh.  This version corrects the comparison operator.

This has been fixed in upstream master branch but the
release/2014.12 branch is frozen so we append it here.

Signed-off-by: Drew Moseley <drew_moseley@mentor.com>