libvpx: fix installation issue due to file permissions
do_install fails if correct permissions are not set for the
installation files as follows:

ERROR: QA Issue: libvpx: /libvpx/usr/lib64/libvpx.so.2.0.0 is
owned by uid 1000, which is the same as the user running bitbake.
This may be due to host contamination [host-user-contaminated]

We fix this by setting appropriate permissions as needed.

Signed-off-by: Awais Belal <awais_belal@mentor.com>