e3000/linux-yocto-rt: upgrade recipe for the RT kernel to v5.4.15
This upgrades the linux-yocto-rt recipe as the previous one does not
have the preempt-rt patch despite being on the rt kernel branch i.e.
"v5.4/standard/preempt-rt/base".

We upgrade to v5.4.15 as this was the first stable release of
the PREEMPT_RT patched kernel in yocto's kernel repo that also
has the corresponding kernel-meta available.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>